fix: Jwt 에러 핸들러 리턴 타입 수정
- 기존 ErrorResponse 타입을 인식하지 못하는 이슈 존재
- Listener 측 Function<I,O> 타입에서 정의한 Output 타입으로 동일하게 매핑
- 유저 Role 부족으로 인한 예외는 별도로 처리하도록 수정
