+++ auth/src/main/java/ddalkak/auth/common/exception/RoleMismatchException.java
@@ -1,0 +1,7 @@
+package ddalkak.auth.common.exception;
+
+public class RoleMismatchException extends RuntimeException {
+    public RoleMismatchException(String message) {
+        super(message);
+    }
+}
