+++ auth/src/main/java/ddalkak/auth/dto/response/ApiGatewayLambdaResponse.java
@@ -1,5 +1,7 @@
 package ddalkak.auth.dto.response;
 
+import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 
 import java.util.HashMap;
@@ -10,7 +12,19 @@ public class ApiGatewayLambdaResponse {
     private boolean isAuthorized;
     private Map<String, Object> context;
 
-    public ApiGatewayLambdaResponse() {
+    @Builder(access = AccessLevel.PRIVATE)
+    public ApiGatewayLambdaResponse(boolean isAuthorized, Map<String, Object> context) {
+        this.isAuthorized = isAuthorized;
+        this.context = context;
+    }
+
+    public static ApiGatewayLambdaResponse errorOf(String message) {
+        ApiGatewayLambdaResponse response = ApiGatewayLambdaResponse.builder()
+                .isAuthorized(false)
+                .context(new HashMap<>())
+                .build();
+        response.addErrorMessage(message);
+        return response;
     }
 
     public ApiGatewayLambdaResponse(boolean isAuthorized) {
@@ -20,5 +34,9 @@ public class ApiGatewayLambdaResponse {
 
     public void addContext(String key, Object value) {
         context.put(key, value);
+    }
+
+    private void addErrorMessage(String message) {
+        this.context.put("message", message);
     }
 }
